refactor: 이벤트 컨슈머 모듈 네임 변경
