+++ app-event-consumer/build.gradle.kts
