+++ app-event-consumer/src/main/java/site/ymango/user/UserCreateConsumer.java
@@ -1,4 +1,4 @@
-package site.ymango;
+package site.ymango.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
