+++ settings.gradle.kts
@@ -1,6 +1,6 @@
 rootProject.name = "flirting"
 
-include(":app-service-api", ":app-point-consumer")
+include(":app-service-api", ":app-event-consumer")
 include(":domain", ":domain:domain-rds", ":domain:domain-redis")
 include(":gateway")
 include(":core")
